Extract PackageLink since we no longer use it but it'll come in handy in the future
